cat file packing agian
On my end this works if the cat files and the content.xml file is in ...\\X4 Foundations\extensions\jonm_shaders
Try that and let me know if it doesn't work!
If the shader files are in the folder or not, does not matter for it to work on my end.