ext_01.cat pathes fixed finally!
ext_01.cat was pointing to the wrong folders, making the include files in the ext_01.dat file not load. I had code files in the same folder that aren't pushed to git, that it read instead making it give a false positive of working. Now tested with code files temporarily removed.